Making the test converter and validator classes as inner classes to the spec
